Refactor lobby model into crate
