+++ micro-game-lobby/Cargo.toml
@@ -1,18 +1,16 @@
 [package]
-name = "micro-game-lobby"
-version = "0.1.0"
 authors = ["terkwood <38859656+Terkwood@users.noreply.github.com>"]
 edition = "2018"
+name = "micro-game-lobby"
+version = "0.1.0"
 
 [dependencies]
-redis_streams = { git = "https://github.com/Terkwood/BUGOUT", version = "0.3.0", rev = "92b8d91abba494737caca525ecb2a4f0d7044dbd" }
-uuid = { version = "0.8.1", "features" = ["v4", "serde"] }
-serde = "1.0.110"
 bincode = "1.2.1"
-serde_derive = "1.0.110"
-log = "0.4.8"
-env_logger = "0.7.1"
+community-redis-streams = {version = "0.1.1", package = "redis-streams"}
 crossbeam = "0.7.3"
 crossbeam-channel = "0.4.2"
-community-redis-streams = { version = "0.1.1", package = "redis-streams" }
-redis = { version = "0.16.0", features = ["r2d2"] }
+env_logger = "0.7.1"
+lobby-model = {path = "lobby-model"}
+log = "0.4.8"
+redis = {version = "0.16.0", features = ["r2d2"]}
+redis_streams = {git = "https://github.com/Terkwood/BUGOUT", version = "0.3.0", rev = "92b8d91abba494737caca525ecb2a4f0d7044dbd"}
