+++ micro-game-lobby/lobby-model/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+authors = ["Terkwood <38859656+Terkwood@users.noreply.github.com>"]
+edition = "2018"
+name = "lobby-model"
+version = "0.1.0"
+
+[dependencies]
+core-model = {git = "https://github.com/Terkwood/BUGOUT", branch = "fix/color-choose-model"}
+serde = "1.0.116"
+serde_derive = "1.0.116"
